feat: VI-13156 Start python gRPC client support
[VI-13156](https://outreach-io.atlassian.net/browse/VI-13156)

Updates `manifest.yaml` to support `python` as an entry in the
`grpcClients` array.  Support for actually building and managing the
release of Python gRPC clients is still a work in progress.  We need
this manifest change to unblock further progress.

Updates `.releaserc.yaml.tpl` to invoke python `build.sh` and
`publish.sh` scripts, intended to be analogous to the same scripts in
Ruby land.  These scripts are not yet implemented in `devbase` and so
these steps will fail.